Move sheets navigator component to its folder
